Fixed overflow during weight calculation
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/biobin@3656 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ trunk/src/biobin/PopulationManager.cpp
@@ -264,7 +264,7 @@ int PopulationManager::getTotalContrib(const Locus& loc) const{
 	return n_var;
 }
 
-float PopulationManager::calcBrowningWeight(int N, int M) const{
+float PopulationManager::calcBrowningWeight(unsigned long N, unsigned long M) const{
 	// Madsen + Browning Weight
 	// A Groupwise Association Test for Rare Mutations Using a Weighted Sum Statistic
 	// PLOSGenetics, Feb 2009, Vol. 5, Issue 2, e10000384
